[#215] Refactor: food 도메인 Service+Facade 패턴으로 마이그레이션
- FoodFinder/FoodManager/FoodModifier/FoodRemover Executor 삭제
- FoodService에 Executor 로직 흡수 (createFood, updateFood, deleteFood 등)
- FoodFacade 신규 생성: S3 업로드와 DB 트랜잭션 분리
- FoodController: createFood/updateFood → FoodFacade 위임
- NotificationProcessor: FoodFinder → FoodRepository 직접 주입으로 전환
- RefrigeratorService: FoodRemover → FoodRepository 직접 주입으로 전환
- Executor 테스트 삭제, FoodServiceIntegrationTest로 케이스 이관
- ControllerTestSupport에 FoodFacade MockitoBean 추가
